Refactor load video logic
