+++ src/video-load/VideoLoader.ts
@@ -1,0 +1,13 @@
+export type VideoFile = {
+  fileName: string
+  sizeBytes: number
+  signature: string
+  frameCount: number
+  frameRate: number
+  orientationDegrees: number
+}
+
+export interface VideoLoader {
+  // TODO Specify return type
+  load(file: File): Promise<any>
+}
